refactor/godClass-docClassUtil
